+++ components/article-view.tsx
@@ -1,6 +1,6 @@
-import { useArticle } from "@pantheon-systems/pcc-react-sdk";
-import type { Article, PublishingLevel } from "@pantheon-systems/pcc-react-sdk";
-import { ArticleRenderer } from "@pantheon-systems/pcc-react-sdk/components";
+import { useArticle } from "@pantheon-systems/cpub-react-sdk";
+import type { Article, PublishingLevel } from "@pantheon-systems/cpub-react-sdk";
+import { ArticleRenderer } from "@pantheon-systems/cpub-react-sdk/components";
 import type { Metadata } from "next";
 import Image from "next/image";
 import Link from "next/link";