+++ components/smart-components/index.ts
@@ -1,7 +1,7 @@
 import {
   ServersideSmartComponentMap,
   SmartComponentMap,
-} from "@pantheon-systems/pcc-react-sdk/components";
+} from "@pantheon-systems/cpub-react-sdk/components";
 import React from "react";
 import { withSmartComponentErrorBoundary } from "./error-boundary";
 import LeadCapture from "./lead-capture";