+++ pages/examples/pagination.tsx
@@ -3,7 +3,7 @@ import {
   ArticleWithoutContent,
   PCCConvenienceFunctions,
   Site,
-} from "@pantheon-systems/pcc-react-sdk";
+} from "@pantheon-systems/cpub-react-sdk";
 import { NextSeo } from "next-seo";
 import { ArticleGrid } from "../../components/grid";
 import Layout from "../../components/layout";