+++ pages/examples/ssg-isr/[...uri].tsx
@@ -1,8 +1,8 @@
 import {
   PCCConvenienceFunctions,
   type Article,
-} from "@pantheon-systems/pcc-react-sdk";
-import { getArticlePathComponentsFromContentStructure } from "@pantheon-systems/pcc-react-sdk/server";
+} from "@pantheon-systems/cpub-react-sdk";
+import { getArticlePathComponentsFromContentStructure } from "@pantheon-systems/cpub-react-sdk/server";
 import { GetStaticPaths, GetStaticProps } from "next";
 import { NextSeo } from "next-seo";
 import { useSearchParams } from "next/navigation";